feat: Add utility functions for MCP query execution and progress updates
- Implemented asynchronous function `run_mcp_query` for executing MCP queries.
- Created `update_progress` function to handle progress updates with logging.
- Developed `process_animation_results` to process and curate animation resources.
- Added `process_documentation_results` for extracting code examples from documentation.
- Implemented `process_frontend_component_results` to extract HTML/CSS/JS from component results.
- Created `process_frontend_templates_results` to curate template resources based on type.
- Developed `process_web_search_results` to format web search results for code generation.
- Added cleanup functions for managing running processes and unused ports.
- Implemented project type detection and environment preparation for various frameworks.
- Created functions for starting, stopping, and restarting application previews.
- Added logging functionality for tracking preview status and errors.
- Implemented dependency installation and virtual environment setup for projects.
