+++ src/mcp/handlers.py
@@ -13,207 +13,11 @@ from src.config.frontend_resources import (
     ANIMATION_RESOURCES,
     BEST_PRACTICES
 )
-
-def process_web_search_results(search_results):
-    """
-    Process web search results into a useful format for code generation.
-    
-    Args:
-        search_results (str): Raw search results from web search tool
-        
-    Returns:
-        str: Processed and formatted search information
-    """
-    # Extract key information from the search results
-    lines = search_results.strip().split('\n')
-    processed_info = []
-    
-    for line in lines:
-        # Skip empty lines
-        if not line.strip():
-            continue
-        
-        # Try to extract URLs if present
-        urls = re.findall(r'https?://[^\s]+', line)
-        if urls:
-            for url in urls:
-                processed_info.append(f"Source: {url}")
-        
-        # Add the line itself
-        processed_info.append(line)
-    
-    return "\n".join(processed_info)
-
-def process_documentation_results(doc_results):
-    """
-    Process documentation search results into code-relevant information.
-    
-    Args:
-        doc_results (str): Raw documentation search results
-        
-    Returns:
-        str: Processed documentation information useful for coding
-    """
-    # Extract code examples if present
-    code_pattern = r'```(?:\w+)?\n(.*?)\n```'
-    code_blocks = re.findall(code_pattern, doc_results, re.DOTALL)
-    
-    processed_info = ["### Documentation Information ###"]
-    processed_info.append(doc_results)
-    
-    if code_blocks:
-        processed_info.append("\n### Extracted Code Examples ###")
-        for i, code in enumerate(code_blocks, 1):
-            processed_info.append(f"Example {i}:\n{code}")
-    
-    return "\n".join(processed_info)
-
-def process_frontend_component_results(component_results):
-    """
-    Process frontend component search results to extract usable code.
-    
-    Args:
-        component_results (str): Raw component search results
-        
-    Returns:
-        str: Processed component information with code examples
-    """
-    # Extract HTML/CSS/JS code
-    html_pattern = r'```(?:html)?\n(.*?)\n```'
-    css_pattern = r'```(?:css)?\n(.*?)\n```'
-    js_pattern = r'```(?:javascript|js)?\n(.*?)\n```'
-    
-    html_blocks = re.findall(html_pattern, component_results, re.DOTALL)
-    css_blocks = re.findall(css_pattern, component_results, re.DOTALL)
-    js_blocks = re.findall(js_pattern, component_results, re.DOTALL)
-    
-    processed_info = ["### Component Information ###"]
-    processed_info.append(component_results)
-    
-    if html_blocks or css_blocks or js_blocks:
-        processed_info.append("\n### Extracted Component Code ###")
-        
-        if html_blocks:
-            processed_info.append("HTML:")
-            for html in html_blocks:
-                processed_info.append(f"{html}")
-        
-        if css_blocks:
-            processed_info.append("\nCSS:")
-            for css in css_blocks:
-                processed_info.append(f"{css}")
-        
-        if js_blocks:
-            processed_info.append("\nJavaScript:")
-            for js in js_blocks:
-                processed_info.append(f"{js}")
-    
-    return "\n".join(processed_info)
-
-def process_frontend_templates_results(template_results, template_type):
-    """
-    Process frontend template search results and add curated resources.
-    
-    Args:
-        template_results (str): Raw template search results
-        template_type (str): Type of template requested
-        
-    Returns:
-        str: Processed template information with links to curated resources
-    """
-    processed_info = ["### Template Resources ###"]
-    processed_info.append(template_results)
-    
-    # Add curated resources
-    template_links = get_templates_by_type(template_type)
-    if template_links:
-        processed_info.append("\n### Curated Template Resources ###")
-        processed_info.append(f"Template type: {template_type}")
-        processed_info.append("Recommended resources:")
-        for i, link in enumerate(template_links, 1):
-            processed_info.append(f"{i}. {link}")
-    
-    return "\n".join(processed_info)
-
-def process_animation_results(animation_results, animation_type):
-    """
-    Process animation resource search results and add curated resources.
-    
-    Args:
-        animation_results (str): Raw animation search results
-        animation_type (str): Type of animation requested
-        
-    Returns:
-        str: Processed animation information with links to curated resources
-    """
-    processed_info = ["### Animation Resources ###"]
-    processed_info.append(animation_results)
-    
-    # Add curated animation resources
-    processed_info.append("\n### Curated Animation Libraries ###")
-    for name, info in ANIMATION_RESOURCES.items():
-        processed_info.append(f"\n{info['name']} - {info['description']}")
-        processed_info.append(f"URL: {info['url']}")
-        processed_info.append(f"CDN: {info['cdn']}")
-    
-    # Add code examples based on animation type
-    if animation_type.lower() in ['hover', 'mouse']:
-        processed_info.append("\n### Hover.css Example ###")
-        processed_info.append("""
-HTML:
-```html
-<link href="https://cdnjs.cloudflare.com/ajax/libs/hover.css/2.3.1/css/hover-min.css" rel="stylesheet">
-<button class="hvr-grow">Hover to Grow</button>
-<div class="hvr-float">Hover to Float</div>
-```
-
-CSS:
-```css
-/* Add to existing CSS if not using the CDN */
-.hvr-grow {
-  display: inline-block;
-  vertical-align: middle;
-  transform: translateZ(0);
-  box-shadow: 0 0 1px rgba(0, 0, 0, 0);
-  backface-visibility: hidden;
-  -moz-osx-font-smoothing: grayscale;
-  transition-duration: 0.3s;
-  transition-property: transform;
-}
-.hvr-grow:hover,
-.hvr-grow:focus,
-.hvr-grow:active {
-  transform: scale(1.1);
-}
-```
-        """)
-    elif animation_type.lower() in ['scroll', 'scrolling']:
-        processed_info.append("\n### AOS (Animate On Scroll) Example ###")
-        processed_info.append("""
-HTML:
-```html
-<link href="https://unpkg.com/aos@2.3.1/dist/aos.css" rel="stylesheet">
-<script src="https://unpkg.com/aos@2.3.1/dist/aos.js"></script>
-
-<div data-aos="fade-up">Fade up on scroll</div>
-<div data-aos="fade-down">Fade down on scroll</div>
-<div data-aos="zoom-in" data-aos-delay="100">Zoom in with delay</div>
-```
-
-JavaScript:
-```javascript
-// Initialize AOS
-document.addEventListener('DOMContentLoaded', function() {
-  AOS.init({
-    duration: 800,
-    easing: 'ease-in-out',
-    once: true
-  });
-});
-```
-        """)
-    
-    return "\n".join(processed_info)
+from src.mcp.steps.process_web_search_results import process_web_search_results
+from src.mcp.steps.process_documentation_results import process_documentation_results
+from src.mcp.steps.process_frontend_component_results import process_frontend_component_results
+from src.mcp.steps.process_frontend_templates_results import process_frontend_templates_results
+from src.mcp.steps.process_animation_results import process_animation_results
 
 def handle_tool_results(tool_name, tool_results, tool_args=None):
     """
