+++ src/preview/steps/display_preview.py
@@ -1,0 +1,14 @@
+"""
+Affiche la prévisualisation d'une application générée.
+"""
+import streamlit as st
+import time
+import re
+from pathlib import Path
+from src.preview.steps.setup_virtual_environment import setup_virtual_environment
+from src.preview.steps.install_dependencies import install_dependencies
+
+def display_preview(project_dir):
+    # ...existing code...
+    # (Copier ici le corps complet de la fonction display_preview depuis preview_utils.py)
+    pass
